+++ packages/browser/__tests__/login/oidc/redirectHandler/AuthCodeRedirectHandler.spec.ts
@@ -38,33 +38,8 @@ import {
 } from "../../../../src/login/oidc/redirectHandler/AuthCodeRedirectHandler";
 import { RedirectorMock } from "../../../../src/login/oidc/__mocks__/Redirector";
 import { SessionInfoManagerMock } from "../../../../src/sessionInfo/__mocks__/SessionInfoManager";
-import { KEY_CURRENT_ISSUER } from "../../../../dist/constant";
-
-class LocalStorageMock {
-  public store: {
-    [key: string]: string;
-  };
-
-  constructor() {
-    this.store = {};
-  }
-
-  public clear() {
-    this.store = {};
-  }
-
-  public getItem(key: string) {
-    return this.store[key] || undefined;
-  }
-
-  public setItem(key: string, value: string) {
-    this.store[key] = value.toString();
-  }
-
-  public removeItem(key: string) {
-    delete this.store[key];
-  }
-}
+import { KEY_CURRENT_SESSION } from "../../../../src/constant";
+import { LocalStorageMock } from "../../../../src/storage/__mocks__/LocalStorage";
 
 const mockJwk = (): JSONWebKey => {
   return {
@@ -383,8 +358,8 @@ describe("AuthCodeRedirectHandler", () => {
 
       // Also check that our issuer was stored correctly (specifically in
       // 'localStorage').
-      expect(window.localStorage.getItem(KEY_CURRENT_ISSUER)).toEqual(
-        testIssuer
+      expect(window.localStorage.getItem(KEY_CURRENT_SESSION)).toEqual(
+        "mySession"
       );
     });
 
@@ -430,11 +405,10 @@ describe("AuthCodeRedirectHandler", () => {
         // @ts-ignore
         header
       ).toMatch(/^DPoP .+$/);
-
       // Also check that our issuer was stored correctly (specifically in
       // 'localStorage').
-      expect(window.localStorage.getItem(KEY_CURRENT_ISSUER)).toEqual(
-        mockIssuer().issuer.toString()
+      expect(window.localStorage.getItem(KEY_CURRENT_SESSION)).toEqual(
+        "mySession"
       );
     });
 
@@ -524,7 +498,9 @@ describe("AuthCodeRedirectHandler", () => {
 
     // Also check that our issuer was stored correctly (specifically in
     // 'localStorage').
-    expect(window.localStorage.getItem(KEY_CURRENT_ISSUER)).toEqual(testIssuer);
+    expect(window.localStorage.getItem(KEY_CURRENT_SESSION)).toEqual(
+      "mySession"
+    );
   });
 
   it("store nothing if the resource server has no session endpoint", async () => {